refactor(webmcp): remove extracted type cycles
