+++ apps/editor/src/ui/editor/state/useEditorViewModel.ts
@@ -60,7 +60,11 @@ import { powerPointIo } from './power-point-io';
 import { textTranslationLayout } from './text-translation-layout';
 import type { TranslationPatch } from './text-translation-layout';
 import { useOperationNotice } from './use-operation-notice';
-import { authoringOperationUiProgress } from './authoringOperationUiProgress';
+import {
+  authoringOperationUiProgress,
+  type DeckTranslationProgressState,
+  type PresentationImportProgressState,
+} from './authoringOperationUiProgress';
 import { useStockMediaLibrary } from './use-stock-media-library';
 import { editorViewModelProgress } from './editorViewModelProgress';
 import { editorViewModelProject } from './editorViewModelProject';
@@ -91,6 +95,10 @@ export type RightPanelTab =
   | 'animations';
 export type TextPreset = 'title' | 'subtitle' | 'body';
 export type { OperationNoticeState } from './use-operation-notice';
+export type {
+  DeckTranslationProgressState,
+  PresentationImportProgressState,
+} from './authoringOperationUiProgress';
 
 export interface MissingPowerPointFont {
   family: string;
@@ -107,20 +115,6 @@ interface PromptPreparationState extends ModelDownloadProgressDetails {
   availability: PromptApiAvailability;
   progress: number;
   status: 'idle' | 'downloading' | 'ready' | 'failed';
-}
-
-export interface PresentationImportProgressState {
-  detail: string;
-  progress: number;
-  stage:
-    | 'reading'
-    | 'inspecting'
-    | 'extracting-objects'
-    | 'downloading-fonts'
-    | 'extracting-media'
-    | 'mapping-animations'
-    | 'opening';
-  title: string;
 }
 
 export type RemoteImportStatus =
@@ -207,13 +201,6 @@ function getMissingPowerPointFonts(project: ProjectDocument, missingFamilies: st
     fonts.set(family.toLowerCase(), current);
   }
   return Array.from(fonts.values()).sort((a, b) => a.family.localeCompare(b.family));
-}
-
-export interface DeckTranslationProgressState {
-  activePageIds: string[];
-  completedPages: number;
-  currentPageName: string;
-  totalPages: number;
 }
 
 export function useEditorViewModel(services: AppServices) {
